% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_change.R
\name{alpha_to_nu}
\alias{alpha_to_nu}
\title{Changes alpha parameter value to nu}
\usage{
alpha_to_nu(p, n, alpha)
}
\arguments{
\item{p}{The number of variables}

\item{n}{The number of samples}

\item{alpha}{Value for alpha parameter}
}
\value{
returns nu
}
\description{
Changes alpha parameter value to nu
}
