% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_change.R
\name{beta_to_delta}
\alias{beta_to_delta}
\title{Changes beta parameter value to delta}
\usage{
beta_to_delta(p, n, nu, beta)
}
\arguments{
\item{p}{The number of variables}

\item{n}{The number of samples}

\item{nu}{Value for nu parameter}

\item{beta}{Value for beta parameter}
}
\value{
returns delta
}
\description{
Changes beta parameter value to delta
}
