% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{data_generator}
\alias{data_generator}
\title{Generates data with a scale free network structure.}
\usage{
data_generator(n, p, d = 0)
}
\arguments{
\item{n}{The number of samples}

\item{p}{The number of variables}

\item{d}{The number of connected variables. Because the function produces scale-free network with maximum of p-1 connections, this values should be between 1 and p.}
}
\value{
Returns the generated data with precision, covariance and adjacency matrices.
}
\description{
Generates data with a scale free network structure.
}
\examples{
library(HMFGraph)
n <- 200
p <- 100
set.seed(42)
generated_data <- data_generator(n=n, p = p, d=p)
}
