% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_change.R
\name{delta_to_beta}
\alias{delta_to_beta}
\title{Changes delta  parameter value to beta}
\usage{
delta_to_beta(p, n, nu, delta)
}
\arguments{
\item{p}{The number of variables}

\item{n}{The number of samples}

\item{nu}{Value for nu parameter}

\item{delta}{Value for delta parameter}
}
\value{
returns beta
}
\description{
Changes delta  parameter value to beta
}
