\name{HMFGraph-package}
\alias{HMFGraph-package}
\alias{HMFGraph}
\docType{package}
\title{
\packageTitle{HMFGraph}
}
\description{
\packageDescription{HMFGraph}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HMFGraph}
\packageIndices{HMFGraph}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{HMFGraph}

Maintainer: \packageMaintainer{HMFGraph}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
