% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMFGraph_GEM.R
\name{HMFGraph_GEM}
\alias{HMFGraph_GEM}
\title{GEM-algorithm for solving MAP-estimate for Bayesian GGM using hierarchical matrix-f prior}
\usage{
HMFGraph_GEM(
  data,
  p = 0,
  n = 0,
  alpha = -1,
  beta = 0.9,
  max_iters = 5000,
  stop_criterion = 10^(-6),
  epsilon1 = 0,
  epsilon2 = 0,
  B = NULL,
  fixed_B = F,
  inter = 500,
  print_t = T
)
}
\arguments{
\item{data}{The data matrix. Variables on columns and samples on rows. n x p}

\item{p}{The number of variables. The value will be derived from the data matrix dimensions if not specified.}

\item{n}{The number of samples. The value will be derived from the data matrix dimensions if not specified.}

\item{alpha}{The value for hyperparameter alpha}

\item{beta}{The value for hyperparameter beta}

\item{max_iters}{The number of maximum iterarions}

\item{stop_criterion}{Stopping criterion for GEM algorithm}

\item{epsilon1}{A shape parameter for Inverse_gamma prior. If a flat prior on logarithmic scale is used, this values should be set to 0.}

\item{epsilon2}{A scale parameter for Inverse_gamma prior. If a flat prior on logarithmic scale is used, this values should be set to 0.}

\item{B}{The initial value for target matrix B. If fixed_B is TRUE, then this value will not change for the whole sampling process.}

\item{fixed_B}{Boolean value. If TRUE, then no prior for target matrix B is used.}

\item{inter}{Interval of iterations when info about algorithm is printed out.}

\item{print_t}{If FALSE, then no text is printed during the function running process.}
}
\value{
Returns the map estimate for precision matrix. Also returns all values that are used for the algorithm.
}
\description{
GEM-algorithm for solving MAP-estimate for Bayesian GGM using hierarchical matrix-f prior
}
\examples{
library(HMFGraph)
n <- 200
p <- 100
set.seed(42)
generated_data <- data_generator(n=n, p = p)
results_HMFGraph_GEM <- HMFGraph_GEM(generated_data$data, alpha = p * 5 / ( p * 5+n), beta=0.9)
}
