% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMFGraph_GEM_FDR_control.R
\name{HMFGraph_GEM_FDR_control}
\alias{HMFGraph_GEM_FDR_control}
\title{Edge selection for GEM-Map estimate based on a target FDR.}
\usage{
HMFGraph_GEM_FDR_control(HMFGraph_GEM_RESULTS, permutation, target_FDR = 0.2)
}
\arguments{
\item{HMFGraph_GEM_RESULTS}{Results from the function HMFGraph_GEM}

\item{permutation}{Results from the function HMFGraph_GEM_permutation}

\item{target_FDR}{The target FDR}
}
\value{
Returns the adjacency matrix, the map estimate, the variance matrix, lower and upper credibility interval point matrices.
}
\description{
Edge selection for GEM-Map estimate based on a target FDR.
}
