% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMFGraph_gibbs_CI.R
\name{HMFGraph_gibbs_CI}
\alias{HMFGraph_gibbs_CI}
\title{Edge selection based on credibility intervals}
\usage{
HMFGraph_gibbs_CI(posterior_samples, CI = 0.9)
}
\arguments{
\item{posterior_samples}{Posterior samples from the HMFGraph_gibbs_sampler -function}

\item{CI}{The credibility interval to be used for selection edges. On defaut the value is 0.9, or 90 % credibility interval}
}
\value{
Returns an adjacency matrix, the posterior mean and median, and lower and upper credibitiry interval points
}
\description{
Edge selection based on credibility intervals
}
