% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMFGraph_gibbs_sampler.R
\name{HMFGraph_gibbs_sampler}
\alias{HMFGraph_gibbs_sampler}
\title{Gibbs sampler for Hierarchical matrix F using Rcpp}
\usage{
HMFGraph_gibbs_sampler(
  data,
  p = 0,
  n = 0,
  alpha = -1,
  beta = 0.9,
  iters = 5000,
  burn_in = 1000,
  epsilon1 = 0,
  epsilon2 = 0,
  B = 0,
  fixed_B = F,
  print_t = T
)
}
\arguments{
\item{data}{The data matrix. Variables on columns and samples on rows. n x p}

\item{p}{The number of variables. The value will be derived from the data matrix dimensions if not specified.}

\item{n}{The number of samples. The value will be derived from the data matrix dimensions if not specified.}

\item{alpha}{The value for hyperparameter alpha}

\item{beta}{The value for hyperparameter beta}

\item{iters}{The number of iterarions}

\item{burn_in}{Burn in length for Gibbs sampling procedure.}

\item{epsilon1}{A shape parameter for Inverse_gamma prior. If a flat prior on logarithmic scale is used, this values should be set to 0.}

\item{epsilon2}{A scale parameter for Inverse_gamma prior. If a flat prior on logarithmic scale is used, this values should be set to 0.}

\item{B}{The initial value for target matrix B. If fixed_B is TRUE, then this value will not change for the whole sampling process.}

\item{fixed_B}{Boolean value. If TRUE, then no prior for target matrix B is used.}

\item{print_t}{If FALSE, no progress bar will be printed.}
}
\value{
Posterior samples of Omega. Also returns nu, delta, alpha and beta values, and the last samples of B and Phi matrices.
}
\description{
Gibbs sampler for Hierarchical matrix F using Rcpp
}
