% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_change.R
\name{nu_to_alpha}
\alias{nu_to_alpha}
\title{Changes nu parameter value to alpha}
\usage{
nu_to_alpha(p, n, nu)
}
\arguments{
\item{p}{The number of variables}

\item{n}{The number of samples}

\item{nu}{Value for nu parameter}
}
\value{
returns alpha
}
\description{
Changes nu parameter value to alpha
}
