% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMFGraph_GEM_CI.R
\name{HMFGraph_GEM_CI}
\alias{HMFGraph_GEM_CI}
\title{Edge selection for GEM-MAP estimate based on a credible interval.}
\usage{
HMFGraph_GEM_CI(HMFGraph_GEM_RESULTS, CI = 0.9)
}
\arguments{
\item{HMFGraph_GEM_RESULTS}{Results from the function HMFGraph_GEM}

\item{CI}{Credible interval. The default value is 0.9, or 90% credible interval}
}
\value{
Returns the adjacency matrix, the map estimate, the variance matrix, lower and upper credible interval point matrices.
}
\description{
Edge selection for GEM-MAP estimate based on a credible interval.
}
\examples{
library(HMFGraph)
n <- 200
p <- 100
set.seed(42)
generated_data <- data_generator(n=n, p = p)
results_HMFGraph_GEM <- HMFGraph_GEM(generated_data$data, alpha = p*5/(p*5+n), beta=0.9)
results_CI <- HMFGraph_GEM_CI(results_HMFGraph_GEM, CI = 0.8)
library(qgraph)
qgraph(results_CI$adjacency_matrix)
}
