% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMFGraph_GEM_FDR_control.R
\name{HMFGraph_GEM_FDR_control}
\alias{HMFGraph_GEM_FDR_control}
\title{Edge selection for GEM-Map estimate based on a target FDR.}
\usage{
HMFGraph_GEM_FDR_control(HMFGraph_GEM_RESULTS, permutation, target_FDR = 0.2)
}
\arguments{
\item{HMFGraph_GEM_RESULTS}{Results from the function HMFGraph_GEM}

\item{permutation}{Results from the function HMFGraph_GEM_permutation}

\item{target_FDR}{The target FDR}
}
\value{
Returns the adjacency matrix, the map estimate, the variance matrix, lower and upper credible interval point matrices.
}
\description{
Edge selection for GEM-Map estimate based on a target FDR.
}
\examples{
library(HMFGraph)
n <- 200
p <- 100
set.seed(42)
generated_data <- data_generator(n=n, p = p)
results_HMFGraph_GEM <- HMFGraph_GEM(generated_data$data, alpha = p * 5 / ( p * 5+n), beta=0.9)
permutations <- HMFGraph_GEM_permutations(generated_data$data, results_HMFGraph_GEM, number_of_permutations = 100, parallel = F)
results_FDR <- HMFGraph_GEM_FDR_control(results_HMFGraph_GEM, permutations, target_FDR = 0.2) library(qgraph)
library(qgraph)
qgraph(results_FDR$adjacency_matrix)
}
