% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMFGraph_GEM_permutations.R
\name{HMFGraph_GEM_permutations}
\alias{HMFGraph_GEM_permutations}
\title{HMFGraph GEM algorithm with permutated data.}
\usage{
HMFGraph_GEM_permutations(
  data,
  HMFGraph_GEM_RESULTS,
  number_of_permutations = 50,
  parallel = FALSE,
  seed = FALSE,
  n_cores = 0
)
}
\arguments{
\item{data}{The original data set}

\item{HMFGraph_GEM_RESULTS}{Results from the function HMFGraph_GEM}

\item{number_of_permutations}{The number of permutations to be run}

\item{parallel}{If TRUE, then parallel computing is used.}

\item{seed}{A seed for permutation function.}

\item{n_cores}{The number of cores for the permutation calculations}
}
\value{
Returns two vectors and vector on quantile points. The first vector consists the number of estimated connections for each quantile point with the original data and the second consist median number of connections for permutated data.
}
\description{
HMFGraph GEM algorithm with permutated data.
}
