% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMFGraph_gibbs_CI.R
\name{HMFGraph_gibbs_CI}
\alias{HMFGraph_gibbs_CI}
\title{Edge selection based on credible intervals}
\usage{
HMFGraph_gibbs_CI(posterior_samples, CI = 0.9)
}
\arguments{
\item{posterior_samples}{Posterior samples from the HMFGraph_gibbs_sampler -function}

\item{CI}{The credible interval to be used for selection edges. On defaut the value is 0.9, or 90 % credible interval}
}
\value{
Returns an adjacency matrix, the posterior mean and median, and lower and upper credible interval points
}
\description{
Edge selection based on credible intervals
}
\examples{
library(HMFGraph)
n <- 200
p <- 100
set.seed(42)
generated_data <- data_generator(n=n, p = p)
results_HMFGraph_gibbs <- HMFGraph_gibbs_sampler(generated_data$data, alpha = p*5/(p*5+n), beta=0.9, iters = 5000, burn_in = 1000)
results_gibbs_CI <- HMFGraph_gibbs_CI(results_HMFGraph_gibbs, CI = 0.8)
library(qgraph)
qgraph(results_gibbs_CI$adjacency_matrix)
}
