% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_change.R
\name{logmarginal}
\alias{logmarginal}
\title{Title}
\usage{
logmarginal(x, n, p, eigen_D, eigen_inv_DS)
}
\arguments{
\item{x}{}

\item{n}{}

\item{p}{}

\item{eigen_D}{}

\item{eigen_inv_DS}{}
}
\description{
Title
}
